+++ train/featuresearch.py
@@ -2,6 +2,7 @@
 
 import argparse
 import concurrent.futures as cf
+import logging
 import os
 import time
 from datetime import timedelta
@@ -13,12 +14,14 @@ from sklearn.model_selection import train_test_split
 from tabulate import tabulate
 from tqdm import tqdm
 
-from .train_model import ModelType, get_model_type
+from .train_model import DEFAULT_MODEL_LOCATION
 from .training_utils import (
     DataVectors,
     evaluate,
     load_datasets,
 )
+
+logger = logging.getLogger(__name__)
 
 DISCARDED_FEATURES = {
     0: [],
@@ -71,7 +74,7 @@ def train_model_feature_search(
     save_model: str,
     seed: int,
     keep_model: bool,
-    model_type: ModelType,
+    combine_name_labels: bool,
 ) -> dict:
     """Train model using selected features returning model performance statistics,
     model parameters and elapsed training time.
@@ -91,8 +94,8 @@ def train_model_feature_search(
         testing sets.
     keep_model : bool
         If True, keep model after evaluation, otherwise delete it.
-    model_type : ModelType
-        Type of model gridsearch is being performed on.
+    combine_name_labels : bool, optional
+        If True, combine all NAME labels into a single NAME label.
 
     Returns
     -------
@@ -140,8 +143,8 @@ def train_model_feature_search(
             "feature.minfreq": 0,
             "feature.possible_states": True,
             "feature.possible_transitions": True,
-            "c1": 0.25,
-            "c2": 0.75,
+            "c1": 0.6,
+            "c2": 0.5,
             "max_linesearch": 5,
             "num_memories": 3,
             "period": 10,
@@ -157,7 +160,7 @@ def train_model_feature_search(
     tagger = pycrfsuite.Tagger()  # type: ignore
     tagger.open(str(save_model_path))
     labels_pred = [tagger.tag(X) for X in features_test]
-    stats = evaluate(labels_pred, truth_test, seed, model_type)
+    stats = evaluate(labels_pred, truth_test, seed, combine_name_labels)
 
     if not keep_model:
         save_model_path.unlink(missing_ok=True)
@@ -180,8 +183,17 @@ def feature_search(args: argparse.Namespace):
         Feature search configuration
     """
     vectors = load_datasets(
-        args.database, args.table, args.datasets, get_model_type(args.model)
+        args.database,
+        args.table,
+        args.datasets,
+        discard_other=True,
+        combine_name_labels=args.combine_name_labels,
     )
+
+    if args.save_model is None:
+        save_model = DEFAULT_MODEL_LOCATION
+    else:
+        save_model = args.save_model
 
     argument_sets = []
     for feature_set in DISCARDED_FEATURES.keys():
@@ -189,23 +201,24 @@ def feature_search(args: argparse.Namespace):
             feature_set,
             vectors,
             args.split,
-            args.save_model,
+            save_model,
             args.seed,
             args.keep_models,
-            get_model_type(args.model),
+            args.combine_name_labels,
         ]
         argument_sets.append(arguments)
 
-    print(f"[INFO] Grid search over {len(argument_sets)} feature sets.")
-    print(f"[INFO] {args.seed} is the random seed used for the train/test split.")
+    logger.info(f"Grid search over {len(argument_sets)} feature sets.")
+    logger.info(f"{args.seed} is the random seed used for the train/test split.")
 
-    eval_results = []
     with cf.ProcessPoolExecutor(max_workers=args.processes) as executor:
         futures = [
             executor.submit(train_model_feature_search, *a) for a in argument_sets
         ]
-        for future in tqdm(cf.as_completed(futures), total=len(futures)):
-            eval_results.append(future.result())
+        eval_results = [
+            future.result()
+            for future in tqdm(cf.as_completed(futures), total=len(futures))
+        ]
 
     # Sort with highest sentence accuracy first
     eval_results = sorted(