+++ train/gridsearch.py
@@ -2,6 +2,7 @@
 
 import argparse
 import concurrent.futures as cf
+import logging
 import os
 import time
 from datetime import timedelta
@@ -14,12 +15,14 @@ from sklearn.model_selection import train_test_split
 from tabulate import tabulate
 from tqdm import tqdm
 
-from .train_model import DEFAULT_MODEL_LOCATION, ModelType, get_model_type
+from .train_model import DEFAULT_MODEL_LOCATION
 from .training_utils import (
     DataVectors,
     evaluate,
     load_datasets,
 )
+
+logger = logging.getLogger(__name__)
 
 # Valid parameter options for LBFGS training algorithm and expected types
 VALID_LBFGS_PARAMS = {
@@ -319,7 +322,11 @@ def generate_argument_sets(args: argparse.Namespace) -> list[list]:
         one of the combinations of algorithms and parameters
     """
     vectors = load_datasets(
-        args.database, args.table, args.datasets, get_model_type(args.model)
+        args.database,
+        args.table,
+        args.datasets,
+        discard_other=True,
+        combine_name_labels=args.combine_name_labels,
     )
 
     # Generate list of arguments for all combinations parameters for each algorithm
@@ -345,7 +352,7 @@ def generate_argument_sets(args: argparse.Namespace) -> list[list]:
             params = params | args.global_params
 
         if args.save_model is None:
-            save_model = DEFAULT_MODEL_LOCATION[args.model]
+            save_model = DEFAULT_MODEL_LOCATION
         else:
             save_model = args.save_model
 
@@ -359,7 +366,7 @@ def generate_argument_sets(args: argparse.Namespace) -> list[list]:
                 save_model,
                 args.seed,
                 args.keep_models,
-                get_model_type(args.model),
+                args.combine_name_labels,
             ]
             argument_sets.append(arguments)
 
@@ -374,7 +381,7 @@ def train_model_grid_search(
     save_model: str,
     seed: int,
     keep_model: bool,
-    model_type: ModelType,
+    combine_name_labels: bool,
 ) -> dict:
     """Train model using given training algorithm and parameters,
     returning model performance statistics, model parameters and elapsed training time.
@@ -396,8 +403,8 @@ def train_model_grid_search(
         testing sets.
     keep_model : bool
         If True, keep model after evaluation, otherwise delete it.
-    model_type : ModelType
-        Type of model gridsearch is being performed on.
+    combine_name_labels : bool, optional
+        If True, combine all NAME labels into a single NAME label.
 
     Returns
     -------
@@ -446,7 +453,7 @@ def train_model_grid_search(
     tagger = pycrfsuite.Tagger()  # type: ignore
     tagger.open(str(save_model_path))
     labels_pred = [tagger.tag(X) for X in features_test]
-    stats = evaluate(labels_pred, truth_test, seed, model_type)
+    stats = evaluate(labels_pred, truth_test, seed, combine_name_labels)
 
     if not keep_model:
         save_model_path.unlink(missing_ok=True)
@@ -489,14 +496,15 @@ def grid_search(args: argparse.Namespace):
 
     arguments = generate_argument_sets(args)
 
-    print(f"[INFO] Grid search over {len(arguments)} hyperparameters combinations.")
-    print(f"[INFO] {args.seed} is the random seed used for the train/test split.")
+    logger.info(f"Grid search over {len(arguments)} hyperparameters combinations.")
+    logger.info(f"{args.seed} is the random seed used for the train/test split.")
 
-    eval_results = []
     with cf.ProcessPoolExecutor(max_workers=args.processes) as executor:
         futures = [executor.submit(train_model_grid_search, *a) for a in arguments]
-        for future in tqdm(cf.as_completed(futures), total=len(futures)):
-            eval_results.append(future.result())
+        eval_results = [
+            future.result()
+            for future in tqdm(cf.as_completed(futures), total=len(futures))
+        ]
 
     # Sort with highest sentence accuracy first
     eval_results = sorted(